+++ src/testing.ts
@@ -3,3 +3,4 @@ export * from './firestore/testing.js';
 export * from './identity-platform/testing.js';
 export * from './pubsub/testing/index.js';
 export * from './spanner/testing.js';
+export * from './testing/index.js';